refactor: 유저 프로필 매치 가드를 프로필 모듈로 이동
