+++ nestjs-BE/server/src/auth/auth.module.ts
@@ -6,7 +6,6 @@ import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
 import { JwtStrategy } from './strategies/jwt.strategy';
 import { JwtAuthGuard } from './guards/jwt-auth.guard';
-import { MatchUserProfileGuard } from './guards/match-user-profile.guard';
 import { IsProfileInSpaceGuard } from './guards/is-profile-in-space.guard';
 import { UsersModule } from '../users/users.module';
 import { ProfilesModule } from '../profiles/profiles.module';
@@ -27,12 +26,10 @@ import { ProfileSpaceModule } from '../profile-space/profile-space.module';
     AuthService,
     JwtStrategy,
     { provide: APP_GUARD, useClass: JwtAuthGuard },
-    MatchUserProfileGuard,
     IsProfileInSpaceGuard,
   ],
   exports: [
     AuthService,
-    MatchUserProfileGuard,
     ProfilesModule,
     IsProfileInSpaceGuard,
     ProfileSpaceModule,
