+++ nestjs-BE/server/src/invite-codes/invite-codes.module.ts
@@ -3,9 +3,10 @@ import { InviteCodesService } from './invite-codes.service';
 import { InviteCodesController } from './invite-codes.controller';
 import { SpacesModule } from '../spaces/spaces.module';
 import { AuthModule } from '../auth/auth.module';
+import { ProfilesModule } from '../profiles/profiles.module';
 
 @Module({
-  imports: [AuthModule, SpacesModule],
+  imports: [AuthModule, SpacesModule, ProfilesModule],
   controllers: [InviteCodesController],
   providers: [InviteCodesService],
 })
